requirements.txt
boto3==1.28.30
botocore==1.31.30
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
idna==3.4
jmespath==1.0.1
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
wrapt==1.15.0
